+++ src/main/java/de/fraunhofer/iosb/ilt/frostclient/models/SensorThingsTaskingV11.java
@@ -80,6 +80,7 @@ public class SensorThingsTaskingV11 implements DataModel {
             throw new IllegalArgumentException("Already initialised.");
         }
         this.mr = modelRegistry;
+        mr.addDataModel(this);
 
         mr.registerEntityType(etActuator);
         mr.registerEntityType(etTask);